runner/requirements.txt
python-dotenv
requests
speedtest-cli